requirements.txt
Django
djangorestframework
pillow
gunicorn==19.1.1
cloudinary 
django-cloudinary-storage